refactor(template): remove unused getDatabaseConfig and inline DB config
Simplify template tests by inlining database config and deleting unused helper.
