+++ template/scenarios/example-01-basic-setup/tests/simple-test.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { getDatabaseConfig } from '../../../tests/database-isolation';
+import { getDatabaseConfig } from '../../../tests/db-config';
 import { validateDatabaseState } from '../../../tests/test-utils';
 
 test.describe('example-01-basic-setup', () => {
